requirements.txt
Flask==2.3.3
PyPDF2==1.27.1
pycryptodome==3.18.0
cryptography==39.0.2
requests==2.31.0
gunicorn==23.0.0
flask-cors==3.0.10
qrcode==7.3.1
Pillow==9.4.0
reportlab==3.6.12
